requirements.txt
openai
langchain
langchain-openai
langchain-community
faiss-cpu
sentence-transformers
pypdf2
pdfplumber
python-docx
fastapi
uvicorn
streamlit
python-dotenv
pydantic
tiktoken
numpy
audio-recorder-streamlit
python-multipart
